requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
defusedxml==0.5.0
Django==2.0
django-bootstrap-pagination==1.7.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.2.0
idna==2.8
oauthlib==3.0.1
Pillow==5.2.0
pkg-resources==0.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
